model/requirements.txt
flask
grpcio
grpcio-tools
scikit-learn==0.22.2.post1
